+++ NEWS.md
@@ -1,6 +1,13 @@
+# SPSP 0.2.0
+
+## New features
+
+1. Add the supports to the SCAD-penalized and MCP-penalized regression model through the `fitfun.SP = SCAD.ncvreg`, `fitfun.SP = MCP.ncvreg`. The previous established methods are `ridge.glmnet`. `lasso.glmnet`,
+`adalasso.glmnet`, and `adalassoCV.glmnet`.
+
 # SPSP 0.1.1
 
-## Major changes
+## patch changes
 
 1. Fix the function corresponding to the `intercept `argument. The intercept argument controls whether to include the intercept, but since the SPSP algorithm estimates the intercept separately. It needs to be dealt with separately.